dynamically rendering by stream ID
